refactor(arch): Phase 10 — final cleanup (100% FSD completion)
- Dead code 삭제: lib/fundamental/cacheTtl, lib/news/cacheTtl
- src/lib/ 5파일 → shared/lib, entities, features 이동 + 테스트 2개 colocation
- src/__tests__/ test utils 3개 → shared/test-utils/ 이동
- src/lib/ 완전 제거, src/__tests__/ 완전 제거
- ESLint legacy-lib 타입/규칙 완전 제거
- 27개 파일 import 갱신

Phase 10 of FSD migration (final cleanup)
