+++ src/entities/api-key/__tests__/actions/deleteApiKeyAction.test.ts
@@ -24,7 +24,7 @@ import { getCurrentUser } from '@/entities/session/lib/getCurrentUser';
 import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
 import { deleteApiKeyAction } from '@/entities/api-key/actions/deleteApiKeyAction';
-import { makeFormData } from '@/__tests__/utils/makeFormData';
+import { makeFormData } from '@/shared/test-utils/makeFormData';
 import type { ApiKeyActionState } from '@/entities/api-key/lib';
 
 const mockGetCurrentUser = getCurrentUser as jest.MockedFunction<
