+++ src/entities/news-article/lib/periodLabels.ts
