+++ src/features/auth-oauth/__tests__/lib/kakao.test.ts
@@ -1,5 +1,5 @@
 import { kakaoOAuthAdapter } from '@/features/auth-oauth/lib/kakao';
-import { jsonResponse } from '@/__tests__/fixtures/jsonResponse';
+import { jsonResponse } from '@/shared/test-utils/jsonResponse';
 
 const ORIGINAL_FETCH = global.fetch;
