+++ src/features/pwa-install/lib/detectPwaEnvironment.ts
