--- src/lib/AGENTS.md
@@ -1,7 +1,0 @@
-# AGENTS.md
-
-When working on this directory, read and follow the instructions in `CLAUDE.md` located in this same directory.
-
-The CLAUDE.md file contains critical guidelines, constraints, and operational procedures that must be followed for all code changes and implementations.
-
-**ALWAYS READ CLAUDE.md FIRST before taking any action.**
