+++ src/shared/test-utils/readBlobText.ts
