+++ src/widgets/fundamental/__tests__/hooks/useFundamentalAnalysis.test.tsx
@@ -13,7 +13,7 @@ import { act, renderHook, waitFor } from '@testing-library/react';
 import type { FundamentalAnalysisResponse } from '@y0ngha/siglens-core';
 import type { ReactNode } from 'react';
 import { renderToString } from 'react-dom/server';
-import { readBlobText } from '@/__tests__/utils/readBlobText';
+import { readBlobText } from '@/shared/test-utils/readBlobText';
 
 jest.mock('@/entities/analysis/actions', () => ({
     submitFundamentalAnalysisAction: jest.fn(),
