+++ src/widgets/news/sections/NewsList.tsx
@@ -3,7 +3,7 @@
 import { useNewsPollingWithInvalidation } from '../hooks/useNewsPollingWithInvalidation';
 import type { NewsDisplayItem } from '@/shared/lib/types';
 import { cn } from '@/shared/lib/cn';
-import { NEWS_LIST_PERIOD_LABEL } from '@/lib/news/periodLabels';
+import { NEWS_LIST_PERIOD_LABEL } from '@/entities/news-article/lib/periodLabels';
 import type { NewsImpact, NewsSentiment } from '@y0ngha/siglens-core';
 import { useState } from 'react';
