+++ src/widgets/options/OptionsStaleDataBanner.tsx
@@ -4,7 +4,7 @@ import {
     ET_MARKET_HOURS_DISPLAY,
     KST_EDT_HOURS_DISPLAY,
     KST_EST_HOURS_DISPLAY,
-} from '@/lib/options/marketHoursDisplay';
+} from '@/entities/options-chain/lib/marketHoursDisplay';
 import { EDT_OFFSET_HOURS, getEasternOffsetHours } from '@/shared/lib/eastern';
 
 /**
@@ -23,7 +23,7 @@ import { EDT_OFFSET_HOURS, getEasternOffsetHours } from '@/shared/lib/eastern';
  * 함께 안내해 사용자가 자기 KST 시계로 바로 환산할 수 있도록 한다.
  *
  * 모든 시간 표기(ET 정규장 + KST 환산 + 현재 EDT/EST 판정)는
- * `@/lib/options/marketHoursDisplay`를 single source of truth로 사용한다 —
+ * `@/entities/options-chain/lib/marketHoursDisplay`를 single source of truth로 사용한다 —
  * 같은 ET/KST 환산 문자열을 쓰는 `OpenInterestChart`의 빈 데이터 안내와
  * 표기가 자동으로 일치한다.
  *
